fix components directory to Game/
