+++ src/components/Game/AudioPlayer/AudioPlayer.scss
@@ -1,4 +1,4 @@
-@import "../../static/css/_theme.scss";
+@import "../../../static/css/_theme.scss";
 
 $firstColor:       $blue;
 $secondColor: $dark-blue;
