+++ src/components/Game/GameInput/GameInput.scss
@@ -1,4 +1,4 @@
-@import "../../static/css/_theme.scss";
+@import "../../../static/css/_theme.scss";
 
 form input {
   box-sizing: border-box;
