+++ src/components/Game/GameText/GameText.scss
@@ -1,4 +1,4 @@
-@import "../../static/css/_theme.scss";
+@import "../../../static/css/_theme.scss";
 
 .game-text {
   margin: 20px 0 10px 0;
