+++ src/components/Game/GameText/GameText.xml
