+++ src/components/Game/RecordPlayer/RecordPlayer.scss
@@ -1,4 +1,4 @@
-@import "../../static/css/_theme.scss";
+@import "../../../static/css/_theme.scss";
 
 $firstColor:       $blue;
 $secondColor: $dark-blue;
