+++ src/components/Game/RecordPlayer/RecordPlayer.xml
